test: update snapshots
Although there are no changes to stories to require updating snapshots, there was an update to the styles of the 'show info' button in the storybook preview window. This required that we update snapshots with this small change. Addon UI altering the preview window is an established anti-pattern that we should be able to remedy when we begin to use addons-docs. Related https://github.com/storybookjs/storybook/issues/6049