feat(table): add extraActions to toolbar overflow menu
You can add other actions to the toolbar overflow menu by using the `extraActions`
prop on the view.toolbar. When clicking on an action it will trigger the
`onApplyExtraActions` callback passing the object used to create the action.
This object must contain an `id` and `itemText`, but it can also contain
`hidden`, `disabled`, `isDelete` items.