feat(table): safari misses drop sometimes
Safari can fire mousemove events with buttons:e before it fires the mouse up event. Need to delay that last mousemove to given mouseup a chance to fire first, otherwise the drop would be canceled. Happens if moving the mouse fast will dropping, which is easy to do.

Also found a test case that failed sometimes because a mouseup was missing in it, so it was in an invalid state for the next step.