fix(table): add column for DnD  in the filter/manage columns headers
When drag and drop is enabled, rows get an extra leading column for the drag and drop handle. The base header gets and extra blank column to line up right. The Filter and Manage Columns headers were missing that extra column, so are one column less than the table body when dtag and drop was enabled, causing them to line up wrong. This just adds that DnD blank column to those header rows as well.